Gave Commands their own Class.
