+++ Src/BootCamp.Chapter/TimeCommand.cs
@@ -5,9 +5,9 @@ using System.Linq;
 
 namespace BootCamp.Chapter
 {
-    public static class Command
+    internal class TimeCommand : ICommand
     {
-        public static IEnumerable<String> CreateTimeReport(string[] command, List<Transaction> transactions)
+        public IEnumerable<string> CreateReport(string[] command, List<Transaction> transactions)
         {
             DateTime[] times = new DateTime[2] { new DateTime(2020, 01, 01, 00, 00, 00), new DateTime(2020, 01, 01, 23, 00, 00) };
 
@@ -31,46 +31,6 @@ namespace BootCamp.Chapter
             }
         }
 
-        public static IEnumerable<String> CreateCityReport(string[] command, List<Transaction> transactions)
-        {
-            const string money = "-money";
-            const string items = "-items";
-            const string min = "-min";
-            const string max = "-max";
-
-            if (command.Length != 3)
-            {
-                throw new InvalidCommandException($"{command[0]} has the wrong amount of parameters.");
-            }
-
-            if (command[1] == items)
-            {
-                if (command[2] == min || command[2] == max)
-                {
-                    return FindItemsMaxOrMin(transactions, command[2]);
-                }
-                else
-                {
-                    throw new InvalidCommandException($"{command[2]} is not a valid parameter.");
-                }
-            }
-            else if (command[1] == money)
-            {
-                if (command[2] == min || command[2] == max)
-                {
-                    return FindMoneyMaxOrMin(transactions, command[2]);
-                }
-                else
-                {
-                    throw new InvalidCommandException($"{command[2]} is not a valid parameter.");
-                }
-            }
-            else
-            {
-                throw new InvalidCommandException($"{command[1]} is not a valid parameter.");
-            }
-        }
-
         private static IEnumerable<String> GroupedByTime(List<Transaction> transactions, DateTime[] times)
         {
             const string topRow = "Hour, Count, Earned";
@@ -84,39 +44,6 @@ namespace BootCamp.Chapter
             ).ToList();
 
             return CreateTabelForSoldByTime(topRow, soldByTime, times[0].Hour, times[1].Hour);
-        }
-
-        private static IEnumerable<String> FindMoneyMaxOrMin(List<Transaction> transactions, string maxOrMin)
-        {
-            /*
-            var groupedcity = from i in transactions
-                              group i by i.City into g
-                              select new { g.Key, Total = g.Sum(i => i.Price) };
-
-            */
-            var groupedcity = transactions.GroupBy(i => i.City)
-                                          .Select(g => new { g.Key, Total = g.Sum(i => i.Price) });
-
-            decimal itemsPerCityMaxOrMin = maxOrMin == "-max" ? groupedcity.Max(g => g.Total) : groupedcity.Min(g => g.Total);
-
-            var toReturn = from i in groupedcity
-                           where i.Total == itemsPerCityMaxOrMin
-                           select i.Key.Trim();
-            return toReturn;
-        }
-
-        private static IEnumerable<String> FindItemsMaxOrMin(List<Transaction> transactions, string maxOrMin)
-        {
-            var groupedcity = from i in transactions
-                              group i.City by i.City into g
-                              select new { g.Key, Count = g.Count() };
-
-            int itemsPerCityMaxOrMin = maxOrMin == "-max" ? groupedcity.Max(g => g.Count) : groupedcity.Min(g => g.Count);
-
-            var toReturn = from i in groupedcity
-                           where i.Count == itemsPerCityMaxOrMin
-                           select i.Key.Trim();
-            return toReturn;
         }
 
         private static IEnumerable<String> CreateTabelForSoldByTime(string topRow, IEnumerable<TimeNumberEarned> soldByTime, int startTime, int EndTime)
@@ -180,6 +107,11 @@ namespace BootCamp.Chapter
             }
 
             return true;
+        }
+
+        public void WriteToCSV(string path, IEnumerable<string> toBeWritten)
+        {
+            ReportsManager.WriteTimeTransaction(path, toBeWritten);
         }
     }
 }
